# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Password Management System - API",
        "title": "MyK3y (Password Management System - API)",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support(Erickson Leynes)",
            "email": "erickson.leynes@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8082",
    "paths": {
        "/api/v1/item": {
            "post": {
                "security": [
                    {
                        "Authentication": []
                    }
                ],
                "description": "Create Item",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Item"
                ],
                "summary": "Create Item",
                "parameters": [
                    {
                        "description": "Create new Item",
                        "name": "createItem",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateItemDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/v1/item/{id}": {
            "get": {
                "security": [
                    {
                        "Authentication": []
                    }
                ],
                "description": "Get Item By ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Item"
                ],
                "summary": "Get Item By ID",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Item search by ID",
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/v1/password": {
            "post": {
                "security": [
                    {
                        "Authentication": []
                    }
                ],
                "description": "Generate Password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Password"
                ],
                "summary": "Generate Password",
                "parameters": [
                    {
                        "description": "Generate new Password",
                        "name": "GeneratePassword",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.GeneratePasswordDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/v1/vault": {
            "post": {
                "security": [
                    {
                        "Authentication": []
                    }
                ],
                "description": "Create Vault",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vault"
                ],
                "summary": "Create Vault",
                "parameters": [
                    {
                        "description": "Create new vault",
                        "name": "createvault",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateVaultDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/v1/vault/{id}": {
            "get": {
                "security": [
                    {
                        "Authentication": []
                    }
                ],
                "description": "Get Vault By ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vault"
                ],
                "summary": "Get Vault By ID",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "vault search by ID",
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "security": [
                    {
                        "Authentication": []
                    }
                ],
                "description": "User Login",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "User Login",
                "parameters": [
                    {
                        "description": "User Login",
                        "name": "UserLogin",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.LoginDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "security": [
                    {
                        "Authentication": []
                    }
                ],
                "description": "User Registration",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "User Registration",
                "parameters": [
                    {
                        "description": "Register new User",
                        "name": "UserRegistration",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UserDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.CreateItemDTO": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "email": {
                    "type": "string"
                },
                "itemname": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                },
                "vaultname": {
                    "type": "string"
                }
            }
        },
        "dto.CreateVaultDTO": {
            "type": "object",
            "properties": {
                "vaultname": {
                    "type": "string"
                }
            }
        },
        "dto.GeneratePasswordDTO": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "isNumbersIncluded": {
                    "type": "boolean"
                },
                "isSymbolsIncluded": {
                    "type": "boolean"
                },
                "isUppercaseIncluded": {
                    "type": "boolean"
                },
                "length": {
                    "type": "integer"
                },
                "passwordType": {
                    "type": "string"
                }
            }
        },
        "dto.LoginDTO": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "minLength": 5
                },
                "password": {
                    "type": "string",
                    "minLength": 8
                }
            }
        },
        "dto.UserDTO": {
            "type": "object",
            "required": [
                "email",
                "password",
                "salt",
                "username"
            ],
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "email": {
                    "type": "string",
                    "minLength": 5
                },
                "password": {
                    "type": "string",
                    "minLength": 8
                },
                "salt": {
                    "type": "string",
                    "minLength": 8
                },
                "updated_at": {
                    "type": "string"
                },
                "username": {
                    "type": "string",
                    "minLength": 5
                }
            }
        },
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "Authentication": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}